done the comments view

+++ database/migrations/2023_03_05_143115_create_reviews_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->longText('content');
             $table->float('stars');
-            $table->float('status'); /* active-1  hidden-0   */
+            $table->integer('status'); /* active-1  hidden-0   */
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade')->onUpdate('cascade');
             $table->foreignId('company_id')->constrained('companies')->onDelete('cascade')->onUpdate('cascade');
             $table->timestamps();
